sd: sync to master-417-43a70e8
